Split up index into separate modules
